# Use SECPKG_ATTR_REMOTE_CERT_CHAIN to retrieve server cert during handshake