Fixes #63321

`SECPKG_ATTR_REMOTE_CERT_CHAIN` can be used before the handshake completes (contrary to `SECPKG_ATTR_REMOTE_CERT_CHAIN_CONTEXT` used until now which led to `SEC_E_INVALID_HANDLE` if called before handshake completion).

Tests were updated to assert that the `remoteCertificate` parameter in `LocalCertificateSelectionCallback` is not null.