@rzikm 2022-02-10
On second thought, I am not 100% sure how the new code behaves if the cert chain is longer than 1 certificate, @wfurt thoughts?